Regenerate the bindings for automatic UUID attribute handling

+++ sources/Interop/Windows/um/dwrite_3/DWRITE_FONT_AXIS_TAG.cs
@@ -8,10 +8,10 @@ namespace TerraFX.Interop
     [NativeTypeName("UINT32")]
     public enum DWRITE_FONT_AXIS_TAG : uint
     {
-        DWRITE_FONT_AXIS_TAG_WEIGHT = (((uint)((byte)((byte)('t'))) << 24) | ((uint)((byte)((byte)('h'))) << 16) | ((uint)((byte)((byte)('g'))) << 8) | (uint)((byte)((byte)('w')))),
-        DWRITE_FONT_AXIS_TAG_WIDTH = (((uint)((byte)((byte)('h'))) << 24) | ((uint)((byte)((byte)('t'))) << 16) | ((uint)((byte)((byte)('d'))) << 8) | (uint)((byte)((byte)('w')))),
-        DWRITE_FONT_AXIS_TAG_SLANT = (((uint)((byte)((byte)('t'))) << 24) | ((uint)((byte)((byte)('n'))) << 16) | ((uint)((byte)((byte)('l'))) << 8) | (uint)((byte)((byte)('s')))),
-        DWRITE_FONT_AXIS_TAG_OPTICAL_SIZE = (((uint)((byte)((byte)('z'))) << 24) | ((uint)((byte)((byte)('s'))) << 16) | ((uint)((byte)((byte)('p'))) << 8) | (uint)((byte)((byte)('o')))),
-        DWRITE_FONT_AXIS_TAG_ITALIC = (((uint)((byte)((byte)('l'))) << 24) | ((uint)((byte)((byte)('a'))) << 16) | ((uint)((byte)((byte)('t'))) << 8) | (uint)((byte)((byte)('i')))),
+        DWRITE_FONT_AXIS_TAG_WEIGHT = (((uint)((byte)('t')) << 24) | ((uint)((byte)('h')) << 16) | ((uint)((byte)('g')) << 8) | (uint)((byte)('w'))),
+        DWRITE_FONT_AXIS_TAG_WIDTH = (((uint)((byte)('h')) << 24) | ((uint)((byte)('t')) << 16) | ((uint)((byte)('d')) << 8) | (uint)((byte)('w'))),
+        DWRITE_FONT_AXIS_TAG_SLANT = (((uint)((byte)('t')) << 24) | ((uint)((byte)('n')) << 16) | ((uint)((byte)('l')) << 8) | (uint)((byte)('s'))),
+        DWRITE_FONT_AXIS_TAG_OPTICAL_SIZE = (((uint)((byte)('z')) << 24) | ((uint)((byte)('s')) << 16) | ((uint)((byte)('p')) << 8) | (uint)((byte)('o'))),
+        DWRITE_FONT_AXIS_TAG_ITALIC = (((uint)((byte)('l')) << 24) | ((uint)((byte)('a')) << 16) | ((uint)((byte)('t')) << 8) | (uint)((byte)('i'))),
     }
 }
